+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/reference/GenomeInterval.java
@@ -3,8 +3,8 @@ package de.charite.compbio.jannovar.reference;
 import java.io.Serializable;
 
 import de.charite.compbio.jannovar.Immutable;
+import de.charite.compbio.jannovar.data.ReferenceDictionary;
 import de.charite.compbio.jannovar.impl.util.StringUtil;
-import de.charite.compbio.jannovar.io.ReferenceDictionary;
 
 /**
  * Representation of a genomic interval (chromsome, begin, end).