+++ jannovar-core/src/test/java/de/charite/compbio/jannovar/reference/TranscriptIntervalTest.java
@@ -4,7 +4,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import de.charite.compbio.jannovar.io.ReferenceDictionary;
+import de.charite.compbio.jannovar.data.ReferenceDictionary;
 
 public class TranscriptIntervalTest {