+++ jannovar-filter/src/main/java/de/charite/compbio/jannovar/filter/Gene.java
@@ -2,7 +2,7 @@ package de.charite.compbio.jannovar.filter;
 
 import com.google.common.collect.ImmutableList;
 
-import de.charite.compbio.jannovar.io.ReferenceDictionary;
+import de.charite.compbio.jannovar.data.ReferenceDictionary;
 import de.charite.compbio.jannovar.reference.GenomeInterval;
 import de.charite.compbio.jannovar.reference.Strand;
 import de.charite.compbio.jannovar.reference.TranscriptModel;