+++ jannovar-filter/src/main/java/de/charite/compbio/jannovar/filter/GeneBuilder.java
@@ -8,7 +8,7 @@ import org.slf4j.LoggerFactory;
 import com.google.common.collect.ImmutableList;
 
 import de.charite.compbio.jannovar.JannovarFilterCommandLineParser;
-import de.charite.compbio.jannovar.io.ReferenceDictionary;
+import de.charite.compbio.jannovar.data.ReferenceDictionary;
 import de.charite.compbio.jannovar.reference.GenomeInterval;
 import de.charite.compbio.jannovar.reference.TranscriptModel;