+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 
 ### jannovar-core
 
+* Renaming package `de.charite.compbio.jannovar.io` to `de.charite.compbio.jannovar.data`
 * Renaming `AnnotationLocation.toHGVSString` to `.toHGVSChunk`.
 * Adding `Pedigree.subsetOfMembers`
 * Renaming `GenomeChange` to `GenomeVariant`, same with types having the same